Fitur 3 belum

+++ src/main/java/com/apap/tugas1/model/JabatanPegawaiModel.java
@@ -10,7 +10,7 @@ import javax.persistence.*;
 public class JabatanPegawaiModel implements Serializable{
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private BigInteger id;
+	private Long id;
 	
 	/*@ManyToOne(fetch = FetchType.LAZY, cascade = {CascadeType.PERSIST, CascadeType.MERGE})
 	@JoinColumn(name = "id_pegawai",referencedColumnName = "id", nullable = false)
@@ -40,7 +40,7 @@ public class JabatanPegawaiModel implements Serializable{
 		return id;
 	}*/
 
-	public void setId(BigInteger id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 }
